Add argument parsing, refactor masquerading

+++ cpp/logging.hpp
@@ -16,7 +16,7 @@ void tcp(const std::string& title, const nets::IPv4Packet& packet)
     std::lock_guard lock(logging_lock);
 
     if (!title.empty())
-        std::cout << title << " origin " << static_cast<int>(packet.origin_id) << '\n';
+        std::cout << title << " origin " << packet.origin_id << '\n';
     std::cout << "From " << packet.source_addr() << ":" << packet.tcp_sport()
               << " to " << packet.destination_addr() << ":" << packet.tcp_dport() << '\n';
     std::cout << "SYN " << static_cast<int>(packet.flag_syn())
@@ -32,7 +32,7 @@ void ip(const std::string& title, const nets::IPv4Packet& packet)
     std::lock_guard lock(logging_lock);
 
     if (!title.empty())
-        std::cout << title << " origin " << static_cast<int>(packet.origin_id) << '\n';
+        std::cout << title << " origin " << packet.origin_id << '\n';
     std::cout << "From " << packet.source_addr()
               << " to " << packet.destination_addr() << '\n';
     std::cout << "TTL " << static_cast<int>(packet.ttl()) << " Proto " << static_cast<int>(packet.protocol()) << '\n';
